O7E-50 Renamed VehicleSummary to VehicleOwnerSummary component and update imports in new-vehicle.vue; renamed Page2.vue to VehicleSummary
